torch float instead of str
Former-commit-id: 2b2de14f4df558b41797e9268a6d679adbd29cd1
Former-commit-id: f20f92d2702ef068f19c2373335854703b66b3f3

+++ tests/test_main.py
@@ -1,6 +1,7 @@
 from context import models
 import muon as mu
 import anndata as ad
+import torch
 import numpy as np
 
 n_cells, n_genes, n_peaks = 20, 50, 5
@@ -54,7 +55,7 @@ def test_custom_params():
         mdata,
         force_recompute=True,
         normalize_rows=True,
-        dtype="float",
+        dtype=torch.float,
         device="cpu",
     )
